Language.Javascript.Chart

Description

Module for accessing minified chart code (http://www.chartjs.org/). As an example:

import qualified Language.Javascript.Chart as Chart

main = do
    putStrLn $ "Chart version " ++ show Chart.version ++ " source:"
    putStrLn =<< readFile =<< Chart.file Chart.Chart

This package installs data files containing the chart.js sources, which must be available at runtime. If you want to produce an executable with no dependency on associated data files, you can use the file-embed library (https://hackage.haskell.org/package/file-embed):

{-# LANGUAGE TemplateHaskell #-}

import Data.FileEmbed
import qualified Data.ByteString as BS
import qualified Language.Javascript.Chart as Chart
import Language.Haskell.TH.Syntax

main = print chartContents

chartContents :: BS.ByteString
chartContents = $(embedFile =<< runIO (Chart.file Chart.Chart))

Documentation

data Chart

The Chart code to obtain.

Constructors

Chart
The base chart.js system.

ChartBundle
The base system, plus a compliant version of the moment.js time and date library.

ChartCSS
chart.js CSS, usually unnecessary. See here.

version :: Version

The version of Chart provided by this package. Not necessarily the version of this package, but the versions will match in the first three digits.

file :: Chart -> IO FilePath

A local file containing the minified chart.js code for version.